pyproject.toml
[tool.poetry]
name = "pyopennft"
version = "0.1.0"
description = "An open-source Python framework for real-time fMRI neurofeedback training"
authors = ["OpenNFT Team <opennft@gmail.com>"]
license = "GPL-3.0"
readme = "README.md"
keywords = ["neurofeedback", "neurofeedback-training", "opennft"]

classifiers = [
    "Development Status :: 1 - Planning",
    "Intended Audience :: Science/Research",
    "Topic :: Software Development :: Libraries",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: Implementation :: CPython",
]

include = [
    "CHANGELOG.md",
    "CONTRIBUTORS",
    "LICENSE",
]

packages = [
    { include = "opennft" },
]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
appdirs = "^1.4.4"
omegaconf = "^2.1.1"
numpy = ">=1.22.0"
scipy = ">=1.7.3"
pydicom = ">=2.1.2"
nibabel = "^3.2.1"
loguru = "^0.5.3"
watchdog = "^2.1.2"
pooch = "^1.4.0"
PyYAML = "^6.0"
toml = "^0.10.2"
pydantic = "^1.9.0"
python-rtspm = "^0.2.0"

[tool.poetry.dev-dependencies]
pybind11 = "^2.6.1"
pytest = "^6.1.1"
pytest-benchmark = "^3.2.3"
pytest-dependency = "^0.5.1"
pytest-ordering = "^0.6"
Sphinx = "^3.2.1"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
force_sort_within_sections = true
force_to_top = ["typing", "pytest"]
line_length = 120
lines_after_imports = 2
